requirements.txt
flask
re
boto3
pandas
nltk
werkzeug.utils
logging
joblib
torch
transformers